// src/movies/dto/create-movie.dto.ts
export class CreateMovieDto {
    DOCID: string;
    movieId: string;
    movieSeq: string;
    title: string;
    titleEng: string;
    titleOrg: string;
    titleEtc: string;
    prodYear: string;
    nation: string;
    company: string;
    runtime: string;
    rating: string;
    genre: string;
    repRlsDate: string;
    posters: string;
    stlls: string;
}